Added rule for test expression rule.

+++ src/parser/syntax_error.rs
@@ -1,11 +1,11 @@
 
 pub struct SyntaxError {
-    pub position: usize,
+    pub position: u32,
     pub message: String,
 }
 
 impl SyntaxError {
-    pub fn new(position: usize, message: String) -> Self {
+    pub fn new(position: u32, message: String) -> Self {
         SyntaxError { position, message }
     }
 }
